Crate pallet_staking_async_rc_runtime

The Westend runtime. This can be compiled with #[no_std], ready for Wasm.

Modules

api

dynamic_params

Dynamic params that can be adjusted at runtime.

fast_runtime_binary

governance

New governance configurations for the Kusama runtime.

migrations

The runtime migrations per release.

pallet_reward_point_filler

xcm_config

XCM configurations for Westend.

Macros

add_benchmarks

Calls add_benchmark with all configs from define_benchmarks and passes the first two parameters on.

impl_elections_weights

Implements the weight types for the elections module and a specific runtime. This macro should not be called directly; use impl_runtime_weights instead.

list_benchmarks

Calls list_benchmark with all configs from define_benchmarks and passes the first two parameters on.

Structs

AlwaysDisableForSlashGreaterThan

A naive disabling strategy that always disables the offender for any slash more than S severity.

AnnouncementDepositBase

AnnouncementDepositFactor

AssetHubId

AssetHubLocation

BagThresholds

BasicDeposit

BeefySetIdSessionEntries

BlockExecutionWeight

Time to execute an empty block. Calculated by multiplying the Average with 1.0 and adding 0.

BlockWeights

Block weights base values and limits.

BondingDuration

BrokerId

BrokerPalletId

BrokerPot

Burn

ByteDeposit

ConfigDepositBase

CrowdloanId

DataDepositPerByte

DepositBase

DepositFactor

DisablingLimit

DynamicParameterOrigin

Defines what origin can modify which dynamic parameters.

ElectionBounds

We take the top 22500 nominators as electing voters and all of the validators as electable targets. Whilst this is the case, we cannot and shall not increase the size of the validator intentions.

EndingPeriod

EnsureAssetHub

EpochDuration

ExistentialDeposit

ExpectedBlockTime

ExtrinsicBaseWeight

Time to execute a NO-OP extrinsic, for example System::remark. Calculated by multiplying the Average with 1.0 and adding 0.

FriendDepositFactor

HrmpChannelSizeAndCapacityWithSystemRatio

IdentityValidatorIdeOf

IndexDeposit

IsIdentityCall

A type to identify calls to the Identity pallet. These will be filtered to prevent invocation, locking the state of the pallet and preventing further updates to identities and sub-identities. The locked state will be the genesis state of a new system chain and then removed from the Relay Chain.

LeafVersion

LeasePeriod

MaxActiveValidators

MaxAdditionalFields

MaxAgentsToMigrate

Bounding number of agent pot accounts to be migrated in a single block.

MaxApprovals

MaxAuthorities

MaxBackersPerWinner

MaxBalance

MaxControllersInDeprecationBatch

MaxElectingVoters

MaxExposurePageSize

MaxFriends

MaxHistoricalRevenue

MaxKeys

MaxLocks

MaxMemoLength

MaxNominations

MaxOffenceBatchSize

each offence is 74 bytes max, sending 50 at a time will be 3,700 bytes.

MaxPeerInHeartbeats

MaxPending

MaxProxies

MaxRegistrars

MaxReserves

MaxScheduledPerBlock

MaxSetIdSessionEntries

MaxSignatories

MaxSubAccounts

MaxTemporarySlotPerLeasePeriod

MaxWinnersPerPage

MaxXcmTransactWeight

MaximumSchedulerWeight

MaybeUsePreviousValidatorsElse

A testing type that implements SessionManager, it receives a new validator set from StakingAhClient, but it prevents them from being passed over to the session pallet and just uses the previous session keys.

MbmServiceWeight

MessageProcessor

Message processor to handle any messages that were enqueued into the MessageQueue pallet.

MessageQueueHeapSize

MessageQueueMaxStale

MessageQueueServiceWeight

Amount of weight that can be spent per block to service messages.

MigrationMaxKeyLen

MigrationSignedDepositBase

MigrationSignedDepositPerItem

MinContribution

MinVestedTransfer

MinimumPeriod

NoPreimagePostponement

NposCompactSolution16

NposSolutionPriority

OffchainRepeat

OffchainSolutionLengthLimit

A limit for off-chain phragmen unsigned solution length.

OffchainSolutionWeightLimit

A limit for off-chain phragmen unsigned solution submission.

Offset

OnChainSeqPhragmen

OnDemandPalletId

OnDemandTrafficDefaultValue

OperationalFeeMultiplier

This value increases the priority of Operational transactions by adding a “virtual tip” that’s equal to the OperationalFeeMultiplier * final_fee.

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

ParaDeposit

ParaHeadsRootProvider

A BEEFY data provider that merkelizes all the parachain heads at the current block (sorted by their parachain id).

ParasUnsignedPriority

ParityDbWeight

ParityDB can be enabled with a feature flag, but is still experimental. These weights are available for brave runtime engineers who may want to try this out as default.

PayoutSpendPeriod

Period

PermanentSlotLeasePeriodLength

PreimageBaseDeposit

PreimageByteDeposit

PreimageHoldReason

ProxyDepositBase

ProxyDepositFactor

QueuedOffenceToXcm

RecoveryDeposit

RegistrarDataDepositPerByte

RemoveKeysLimit

ReportLongevity

RocksDbWeight

By default, Substrate uses RocksDB, so this will be the weight used throughout the runtime.

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

RuntimeGenesisConfig

RuntimeOrigin

The runtime origin type representing the origin of a call.

SS58Prefix

SampleLength

SessionKeys

SessionReportToXcm

SessionsPerEra

SignedDepositByte

SignedDepositIncreaseFactor

SignedFixedDeposit

SignedMaxRefunds

SignedMaxSubmissions

SignedPhase

SignedRewardBase

SlashDeferDuration

SpendPeriod

StakingXcmToAssetHub

SubAccountDeposit

SubmissionDeposit

SwapLeases

TemporarySlotLeasePeriodLength

TipCountdown

TipFindersFee

TipReportDepositBase

TransactionByteFee

TreasuryInteriorLocation

TreasuryPalletId

UnsignedPhase

UnvestedFundsAllowedWithdrawReasons

UsePreviousValidators

UsernameDeposit

Version

Enums

BalancesCall

Contains a variant per dispatchable extrinsic that this pallet has.

OriginCaller

ProxyType

The type used to represent the kinds of proxying allowed.

RuntimeCall

The aggregated runtime call type.

RuntimeError

RuntimeEvent

RuntimeFreezeReason

A reason for placing a freeze on funds.

RuntimeHoldReason

A reason for placing a hold on funds.

RuntimeLockId

An identifier for each lock placed on funds.

RuntimeSlashReason

A reason for slashing funds.

RuntimeTask

An aggregation of all Task enums across all pallets included in the current runtime.

RuntimeViewFunction

Runtime query type.

SystemCall

Contains a variant per dispatchable extrinsic that this pallet has.

TimestampCall

Contains a variant per dispatchable extrinsic that this pallet has.

Constants

AVERAGE_ON_INITIALIZE_RATIO

We assume that an on-initialize consumes 1% of the weight on average, hence a single extrinsic will not be allowed to consume more than AvailableBlockRatio - 1%.

BABE_GENESIS_EPOCH_CONFIG

The BABE epoch configuration at genesis.

MAXIMUM_BLOCK_WEIGHT

We allow for 2 seconds of compute with a 6 second average block time. The storage proof size is not limited so far.

NORMAL_DISPATCH_RATIO

We allow Normal extrinsics to fill up the block up to 75%, the rest can be used by Operational extrinsics.

RUNTIME_API_VERSIONS

VERSION

Runtime version (Westend).

WASM_BINARY

WASM_BINARY_BLOATY

WASM_BINARY_PATH

Traits

BuildStorage

Complex storage builder stuff.

Functions

native_version

Native version.

Type Aliases

Address

The address format for describing accounts.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

AssetRate

AssignedSlots

AssignedSlotsConfig

Auctions

AuthorityDiscovery

AuthorityDiscoveryConfig

Authorship

Babe

BabeConfig

Balances

BalancesConfig

Beefy

BeefyConfig

BeefyMmrLeaf

Block

Block type as expected by this runtime.

BlockId

BlockId type as expected by this runtime.

Configuration

ConfigurationConfig

ConvictionVoting

Coretime

CoretimeAssignmentProvider

Crowdloan

Dmp

ElectionProviderMultiPhase

Executive

Executive: handles dispatch to the various modules.

Grandpa

GrandpaConfig

Header

Block header type as expected by this runtime.

Historical

Hrmp

HrmpConfig

Identity

IdentityMigrator

Indices

IndicesConfig

Initializer

MessageQueue

Migrations

All migrations that will run on the next runtime upgrade.

Mmr

MultiBlockMigrations

Multisig

Offences

OnDemandAssignmentProvider

Origins

ParaInclusion

ParaInherent

ParaScheduler

ParaSessionInfo

ParachainsOrigin

Parameters

Paras

ParasConfig

ParasDisputes

ParasShared

ParasSlashing

ParasSudoWrapper

Preimage

PresetStore

PresetStoreConfig

Proxy

Recovery

Referenda

Registrar

RegistrarConfig

RewardPointFiller

RootOffences

RootTesting

Scheduler

Session

SessionConfig

SignedBlock

A Block signed with a Justification

SignedPayload

The payload being signed in transactions.

Slots

Staking

StakingAhClient

StakingAhClientConfig

StakingConfig

Sudo

SudoConfig

System

SystemConfig

Timestamp

TransactionPayment

TransactionPaymentConfig

Treasury

TreasuryConfig

TxExtension

The extension to the basic transaction logic.

UncheckedExtrinsic

Unchecked extrinsic type as expected by this runtime.

UncheckedSignaturePayload

Unchecked signature payload type as expected by this runtime.

Utility

Vesting

VestingConfig

VoterList

Whitelist

XcmPallet

XcmPalletConfig